+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `WalletEvent::SpentOutput` boxed;
 - Allow null nft ids in `prepare_output()` for minting;
 - `TagFeature` & `MetadataFeature` encoded as hex strings instead of utf-8 strings;
+- Call `try_get_outputs_metadata` instead of `try_get_outputs` when possible (more efficient);
 
 ### Removed